idk what to write honestly

+++ src/main/java/io/github/skulli73/main/messageComponents/StartElectionMessageComponent.java
@@ -1,10 +1,13 @@
 package io.github.skulli73.main.messageComponents;
 
 import io.github.skulli73.main.objects.Election;
+import org.javacord.api.entity.channel.PrivateChannel;
 import org.javacord.api.entity.channel.TextChannel;
 import org.javacord.api.entity.message.MessageBuilder;
 import org.javacord.api.entity.message.component.ActionRow;
 import org.javacord.api.entity.message.component.Button;
+import org.javacord.api.entity.permission.Role;
+import org.javacord.api.entity.user.User;
 import org.javacord.api.interaction.MessageComponentInteraction;
 
 import javax.xml.soap.Text;
@@ -21,25 +24,44 @@ public class StartElectionMessageComponent extends AbstractElectionMessageCompon
     @Override
     void execute(MessageComponentInteraction pInteraction) {
         Election lElection = elections.get(Integer.parseInt(pInteraction.getCustomId().substring(1)));
-        TextChannel lChannel = discordApi.getChannelById(lElection.channel).get().asTextChannel().get();
-        try {
-            discordApi.getMessageById(lElection.message, lChannel).get().delete();
-        } catch (InterruptedException | ExecutionException e) {
-            throw new RuntimeException(e);
-        }
-        try {
-            lElection.message = new MessageBuilder()
-                    .addEmbed(lElection.electoralMethod.genericEmbed(lElection))
-                    .addComponents(
-                            ActionRow.of(
-                                    Button.danger("e" + lElection.id, "End Election")
-                            )
-                    )
-                    .send(lChannel)
-                    .get()
-                    .getId();
-        } catch (InterruptedException | ExecutionException e) {
-            throw new RuntimeException(e);
-        }
+        if(lElection.candidates.size()>0) {
+            TextChannel lChannel = discordApi.getChannelById(lElection.channel).get().asTextChannel().get();
+            try {
+                discordApi.getMessageById(lElection.message, lChannel).get().delete();
+            } catch (InterruptedException | ExecutionException e) {
+                throw new RuntimeException(e);
+            }
+            try {
+                lElection.message = new MessageBuilder()
+                        .addEmbed(lElection.electoralMethod.genericEmbed(lElection))
+                        .addComponents(
+                                ActionRow.of(
+                                        Button.danger("e" + lElection.id, "End Election")
+                                )
+                        )
+                        .send(lChannel)
+                        .get()
+                        .getId();
+            } catch (InterruptedException | ExecutionException e) {
+                throw new RuntimeException(e);
+            }
+
+
+            Role lRole = discordApi.getRoleById(lElection.roleId).get();
+            MessageBuilder lMessageBuilder = lElection.electoralMethod.getBallot(lElection);
+            lMessageBuilder.addEmbed(lElection.electoralMethod.genericEmbed(lElection));
+            lMessageBuilder.addComponents(ActionRow.of(
+                    Button.success("s", "Submit")
+            ));
+            for(User lUser: lRole.getUsers()) {
+                try{
+                    PrivateChannel lDmChannel = lUser.openPrivateChannel().get();
+                    lMessageBuilder.send(lDmChannel);
+                } catch(Exception e) {
+                    pInteraction.getChannel().get().sendMessage("I could not message " + lUser.getMentionTag());
+                }
+            }
+        } else
+            pInteraction.createImmediateResponder().append("You require at least one candidate to start an election.").respond();
     }
 }
